rm pinger usage
