+++ nmproxy/peer/peer.go
@@ -151,7 +151,7 @@ func collectMetricsForServerPeers(server string, peerIDAndAddrMap nm_models.Host
 		if peerIDMap, ok := peerIDAndAddrMap[peer.PublicKey.String()]; ok {
 			metric := metrics.GetMetric(server, peer.PublicKey.String())
 			metric.NodeConnectionStatus = make(map[string]bool)
-			connectionStatus := metrics.PeerConnectionStatus(peer.PublicKey.String())
+			connectionStatus := proxy.PeerConnectionStatus(peer.PublicKey.String())
 			for peerID := range peerIDMap {
 				metric.NodeConnectionStatus[peerID] = connectionStatus
 			}
